requirements.txt
torch
torchvision
flask
pandas
numpy
scikit-learn
pillow